Apply Trick Attack to physical Blue Magic
Physical BLU spells should transfer enmity and consume Trick Attack like
a regular attack or weaponskill. Spells aren't provided a `taChar` param
like weaponskills are, so that needs to change - otherwise we have no
ability to divert the generated enmity.

This is a shitty, minimal patch to proof out this idea. It includes some
demonstrative modifications to the Death Scissors handler. The majority
of spells won't need any of this code, so it'd be best to try isolating
the changes to just physicial BLU spells (or at least just BLU spells in
general).